+++ crates/ferroid/src/futures/mod.rs
@@ -1,7 +1,9 @@
-mod snowflake;
-mod ulid;
 mod sleep_provider;
+mod snowflake;
+#[cfg(feature = "ulid")]
+mod ulid;
 
-pub use snowflake::*;
-pub use ulid::*;
 pub use sleep_provider::*;
+pub use snowflake::*;
+#[cfg(feature = "ulid")]
+pub use ulid::*;
