+++ crates/ferroid/src/generator/mod.rs
@@ -1,7 +1,9 @@
 mod snowflake;
 #[cfg(test)]
 mod tests;
+#[cfg(feature = "ulid")]
 mod ulid;
 
 pub use snowflake::*;
+#[cfg(feature = "ulid")]
 pub use ulid::*;
