+++ crates/ferroid/src/id/mod.rs
@@ -1,7 +1,9 @@
 mod id;
 mod snowflake;
+#[cfg(feature = "ulid")]
 mod ulid;
 
 pub use id::*;
 pub use snowflake::*;
+#[cfg(feature = "ulid")]
 pub use ulid::*;
