+++ crates/ferroid/src/lib.rs
@@ -6,7 +6,10 @@ mod futures;
 mod generator;
 mod id;
 mod mono_clock_native;
+
+#[cfg(feature = "ulid")]
 mod rand;
+#[cfg(feature = "ulid")]
 mod random_native;
 mod runtime;
 mod status;
@@ -20,7 +23,9 @@ pub use crate::futures::*;
 pub use crate::generator::*;
 pub use crate::id::*;
 pub use crate::mono_clock_native::*;
+#[cfg(feature = "ulid")]
 pub use crate::rand::*;
+#[cfg(feature = "ulid")]
 pub use crate::random_native::*;
 #[cfg(any(feature = "async-tokio", feature = "async-smol"))]
 pub use crate::runtime::*;
